Refactored cleanup method of FlowBuilder by checking whether temptmpPathList is null or not



See merge request !157

+++ hydrograph.engine/hydrograph.engine.cascading/src/main/java/hydrograph/engine/cascading/integration/FlowBuilder.java
@@ -47,88 +47,88 @@ import hydrograph.engine.utilities.PlatformHelper;
 @SuppressWarnings("rawtypes")
 public class FlowBuilder {
 
-	private static final Logger LOG = LoggerFactory
-			.getLogger(FlowBuilder.class);
+    private static final Logger LOG = LoggerFactory
+            .getLogger(FlowBuilder.class);
 
-	public FlowBuilder() {
-	}
+    public FlowBuilder() {
+    }
 
-	public void buildFlow(RuntimeContext runtimeContext) {
-		LOG.info("Building flow");
-		JAXBTraversal traversal = runtimeContext.getTraversal();
-		Properties hadoopProps = runtimeContext.getHadoopProperties();
-		HydrographJob hydrographJob = runtimeContext.getHydrographJob();
-		addSerializations(hadoopProps);
+    public void buildFlow(RuntimeContext runtimeContext) {
+        LOG.info("Building flow");
+        JAXBTraversal traversal = runtimeContext.getTraversal();
+        Properties hadoopProps = runtimeContext.getHadoopProperties();
+        HydrographJob hydrographJob = runtimeContext.getHydrographJob();
+        addSerializations(hadoopProps);
 
-		FlowConnector flowConnector = PlatformHelper
-				.getHydrographEngineFlowConnector(hadoopProps);
+        FlowConnector flowConnector = PlatformHelper
+                .getHydrographEngineFlowConnector(hadoopProps);
 
-		Cascade[] cascades = new Cascade[traversal.getFlowsNumber().size()];
-		int batchIndex = -1;
-		String flowName = "flow";
-		for (String batch : traversal.getFlowsNumber()) {
-			batchIndex = batchIndex + 1;
-			FlowContext flowContext = new FlowContext(hydrographJob, traversal,
-					hadoopProps);
-			traverseAndConnect(runtimeContext, flowContext, batch);
-			FlowDef flowDef = flowContext.getFlowDef();
-			if (flowConnector instanceof Hadoop2TezFlowConnector) {
-				flowDef = PlatformHelper
-						.addJarsToClassPathInCaseOfTezExecution(flowDef);
-			}
-			if (flowDef.getSources().size() > 0) {
-				flowContext.getCascadeDef().addFlow(
-						flowConnector.connect(flowDef));
-			}
-			cascades[batchIndex] = (new CascadeConnector().connect(flowContext
-					.getCascadeDef().setName(flowName + "_" + batchIndex)));
-			runtimeContext.setFlowContext(batch, flowContext);
-		}
-		runtimeContext.setCascade(cascades);
+        Cascade[] cascades = new Cascade[traversal.getFlowsNumber().size()];
+        int batchIndex = -1;
+        String flowName = "flow";
+        for (String batch : traversal.getFlowsNumber()) {
+            batchIndex = batchIndex + 1;
+            FlowContext flowContext = new FlowContext(hydrographJob, traversal,
+                    hadoopProps);
+            traverseAndConnect(runtimeContext, flowContext, batch);
+            FlowDef flowDef = flowContext.getFlowDef();
+            if (flowConnector instanceof Hadoop2TezFlowConnector) {
+                flowDef = PlatformHelper
+                        .addJarsToClassPathInCaseOfTezExecution(flowDef);
+            }
+            if (flowDef.getSources().size() > 0) {
+                flowContext.getCascadeDef().addFlow(
+                        flowConnector.connect(flowDef));
+            }
+            cascades[batchIndex] = (new CascadeConnector().connect(flowContext
+                    .getCascadeDef().setName(flowName + "_" + batchIndex)));
+            runtimeContext.setFlowContext(batch, flowContext);
+        }
+        runtimeContext.setCascade(cascades);
 
-	}
+    }
 
-	private void addSerializations(Properties hadoopProps) {
-		TupleSerializationProps.addSerialization(hadoopProps,
-				BigDecimalSerialization.class.getName());
-	}
+    private void addSerializations(Properties hadoopProps) {
+        TupleSerializationProps.addSerialization(hadoopProps,
+                BigDecimalSerialization.class.getName());
+    }
 
-	private void traverseAndConnect(RuntimeContext runtimeContext,
-									FlowContext flowContext, String batch) {
+    private void traverseAndConnect(RuntimeContext runtimeContext,
+                                    FlowContext flowContext, String batch) {
 
-		JAXBTraversal traversal = flowContext.getTraversal();
+        JAXBTraversal traversal = flowContext.getTraversal();
 
-		for (String componentId : traversal.getOrderedComponentsList(batch)) {
-			LOG.info("Building parameters for " + componentId);
+        for (String componentId : traversal.getOrderedComponentsList(batch)) {
+            LOG.info("Building parameters for " + componentId);
 
-			ComponentParameters cp = null;
-			BaseAdapter adapterBase = runtimeContext
-					.getAssemblyGeneratorMap().get(componentId);
-			if (adapterBase instanceof InputAdapterBase) {
-				cp = new ComponentParameterBuilder.Builder(componentId,
-						new ComponentParameters(), flowContext, runtimeContext)
-						.setFlowdef().setJobConf().setSchemaFields().build();
-			} else if (adapterBase instanceof OutputAdapterBase) {
-				cp = new ComponentParameterBuilder.Builder(componentId,
-						new ComponentParameters(), flowContext, runtimeContext)
-						.setFlowdef().setInputPipes().setInputFields().setSchemaFields().build();
-			} else if (adapterBase instanceof StraightPullAdapterBase) {
-				cp = new ComponentParameterBuilder.Builder(componentId,
-						new ComponentParameters(), flowContext, runtimeContext)
-						.setFlowdef().setJobConf().setInputPipes().setSchemaFields()
-						.setInputFields().build();
-			} else if (adapterBase instanceof OperationAdapterBase) {
-				cp = new ComponentParameterBuilder.Builder(componentId,
-						new ComponentParameters(), flowContext, runtimeContext)
-						.setFlowdef().setJobConf().setInputPipes().setUDFPath()
-						.setInputFields().setSchemaFields().build();
-			} else if (adapterBase instanceof CommandAdapterBase) {
-				BaseCommandComponent commandComponent = ((CommandAdapterBase) adapterBase)
-						.getComponent();
-				flowContext.getCascadeDef().addFlow(
-						new ProcessFlow(componentId, commandComponent));
-				continue;
-			}
+            ComponentParameters cp = null;
+            BaseAdapter adapterBase = runtimeContext
+                    .getAssemblyGeneratorMap().get(componentId);
+            if (adapterBase instanceof InputAdapterBase) {
+                cp = new ComponentParameterBuilder.Builder(componentId,
+                        new ComponentParameters(), flowContext, runtimeContext)
+                        .setFlowdef().setJobConf().setSchemaFields().build();
+            } else if (adapterBase instanceof OutputAdapterBase) {
+                cp = new ComponentParameterBuilder.Builder(componentId,
+                        new ComponentParameters(), flowContext, runtimeContext)
+                        .setFlowdef().setInputPipes().setInputFields().setSchemaFields().build();
+            } else if (adapterBase instanceof StraightPullAdapterBase) {
+                cp = new ComponentParameterBuilder.Builder(componentId,
+                        new ComponentParameters(), flowContext, runtimeContext)
+                        .setFlowdef().setJobConf().setInputPipes().setSchemaFields()
+                        .setInputFields().build();
+            } else if (adapterBase instanceof OperationAdapterBase) {
+                cp = new ComponentParameterBuilder.Builder(componentId,
+                        new ComponentParameters(), flowContext, runtimeContext)
+                        .setFlowdef().setJobConf().setInputPipes().setUDFPath()
+                        .setInputFields().setSchemaFields().build();
+            } else if (adapterBase instanceof CommandAdapterBase) {
+                BaseCommandComponent commandComponent = ((CommandAdapterBase) adapterBase)
+                        .getComponent();
+                flowContext.getCascadeDef().addFlow(
+                        new ProcessFlow(componentId, commandComponent));
+                continue;
+            }
 
 //			((ComponentGeneratorBase) assemblyGeneratorBase).createAssembly(cp);
 //			BaseComponent component = ((ComponentGeneratorBase) assemblyGeneratorBase)
@@ -136,39 +136,40 @@ public class FlowBuilder {
 
 //			flowContext.getAssemblies().put(componentId, component);
 
-			((ComponentAdapterBase) adapterBase).createAssembly(cp);
-			BaseComponent component = ((ComponentAdapterBase) adapterBase).getAssembly();
-			flowContext.getAssemblies().put(componentId, component);
+            ((ComponentAdapterBase) adapterBase).createAssembly(cp);
+            BaseComponent component = ((ComponentAdapterBase) adapterBase).getAssembly();
+            flowContext.getAssemblies().put(componentId, component);
 
-			LOG.info("Assembly creation completed for " + componentId);
-		}
-	}
+            LOG.info("Assembly creation completed for " + componentId);
+        }
+    }
 
 
+    public void cleanup(List<String> tmpPathList, RuntimeContext runtimeContext) {
+        if (tmpPathList != null) {
+            for (String tmpPath : tmpPathList) {
 
-	public void cleanup(List<String> tmpPathList,RuntimeContext runtimeContext) {
-		for (String tmpPath : tmpPathList) {
+                Path fullPath = new Path(tmpPath);
+                // do not delete the root directory
+                if (fullPath.depth() == 0)
+                    continue;
+                FileSystem fileSystem;
 
-			Path fullPath = new Path(tmpPath);
-			// do not delete the root directory
-			if (fullPath.depth() == 0)
-				continue;
-			FileSystem fileSystem;
+                LOG.info("Deleting temp path:" + tmpPath);
+                try {
+                    fileSystem = FileSystem.get(runtimeContext.getJobConf());
 
-			LOG.info("Deleting temp path:" + tmpPath);
-			try {
-				fileSystem = FileSystem.get(runtimeContext.getJobConf());
+                    fileSystem.delete(fullPath, true);
+                } catch (NullPointerException exception) {
+                    // hack to get around npe thrown when fs reaches root directory
+                    // if (!(fileSystem instanceof NativeS3FileSystem))
+                    throw new RuntimeException(exception);
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
 
-				fileSystem.delete(fullPath, true);
-			} catch (NullPointerException exception) {
-				// hack to get around npe thrown when fs reaches root directory
-				// if (!(fileSystem instanceof NativeS3FileSystem))
-				throw new RuntimeException(exception);
-			} catch (IOException e) {
-				throw new RuntimeException(e);
-			}
-
-		}
-	}
+            }
+        }
+    }
 
 }
